backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
PyGithub==2.9.1
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
